HGMAP/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
scikit-learn>=0.24.2
xgboost>=1.4.2
lightgbm>=3.2.1
catboost>=0.26
optuna>=2.9.1
shap>=0.39.0
lime>=0.2.0
captum>=0.4.0
scipy>=1.7.0
networkx>=2.6.2
skbio>=0.5.6
umap-learn>=0.5.1
boruta>=0.3
